working through highlighting parts of a URL which match.
Ideally, URLPattern should have a simplify() method which renders the value out much like the admindocs does. But that requires me to get upstreamed the _removal_ of some of the admindocs stuff in favour of the actually-correct normalize, because I know for a fact admindocs chokes on some valid url patterns.

Upside if I _can_ get all that changed, I'll be able to make a stronger case that there should just be a gorram management command built in to list all routes/urls in the app, because I'll have consolidated everything ...